feat(25): refactor component folder structure
