+++ src/components/dynamicComponents/carousel/carousel1/carousel1.module.css
