+++ src/components/dynamicComponents/carousel/carousel1/index.tsx
