+++ src/components/dynamicComponents/carousel/index.ts
