+++ src/components/dynamicComponents/footer/footer1/footer1.module.css
