+++ src/components/dynamicComponents/footer/index.ts
