+++ src/components/dynamicComponents/fullSizeBanner/FullSizeBanner.tsx
