+++ src/components/dynamicComponents/fullSizeBanner/animations.ts
