+++ src/components/dynamicComponents/fullSizeBanner/style.ts
