+++ src/components/dynamicComponents/productCard/card1/Card1.module.css
