+++ src/components/dynamicComponents/productCard/card1/index.tsx
