+++ src/components/dynamicComponents/productCard/index.ts
