+++ src/components/dynamicComponents/tiles/common/card.module.css
