+++ src/components/dynamicComponents/tiles/index.ts
