+++ src/components/generic/paletteDisplay/PaletteDisplay.tsx
