+++ src/components/generic/paletteDisplay/index.ts
