--- src/components/userLandingPage/UserLandingPage.tsx
@@ -1,5 +1,0 @@
-function UserLandingPage() {
-  return <p>Home</p>
-}
-
-export default UserLandingPage
